Office Id Implementation and Mini Rules API (#20)
* Search Results for Bombay, Delhi, Benzee Ids
* Implementation of new minirules API (#19)

---------

+++ app/utils/AmadeusSessionManager.java
@@ -2,14 +2,16 @@ package utils;
 
 import com.amadeus.xml.ws._2009._01.wbs_session_2_0.Session;
 import com.compassites.GDSWrapper.amadeus.ServiceHandler;
-import com.compassites.GDSWrapper.amadeus.SessionHandler;
 import com.compassites.constants.AmadeusConstants;
 import models.AmadeusSessionWrapper;
+import models.FlightSearchOffice;
 import org.joda.time.DateTime;
 import org.joda.time.Period;
 import org.joda.time.PeriodType;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import services.AmadeusSourceOfficeService;
 
 import javax.xml.ws.Holder;
 import java.util.Date;
@@ -22,17 +24,92 @@ import java.util.UUID;
 @Service
 public class AmadeusSessionManager {
 
-//    private static List<AmadeusSessionWrapper> sessionList = new ArrayList<>();
     static org.slf4j.Logger logger = LoggerFactory.getLogger("gds");
 
-    /*private ServiceHandler serviceHandler;
+    @Autowired
+    private ServiceHandler serviceHandler;
 
     @Autowired
-    public AmadeusSessionManager(ServiceHandler serviceHandler) {
-        this.serviceHandler = serviceHandler;
-    }*/
+    private AmadeusSourceOfficeService sourceOfficeService;
 
-    public AmadeusSessionWrapper getSession() throws InterruptedException {
+//    private static final ConcurrentHashMap<String, BlockingQueue<AmadeusSessionWrapper>> officeSessionMap = new ConcurrentHashMap<>();
+
+//    public AmadeusSessionWrapper getSession(String officeId) throws Exception {
+//        logger.debug("AmadeusSessionManager getSession called");
+//        // Check if there is a queue for the specified officeId
+//        List<AmadeusSessionWrapper> amadeusSessionWrapperList = AmadeusSessionWrapper.findAllInactiveContextList();
+//        BlockingQueue<AmadeusSessionWrapper> sessionQueue = officeSessionMap.get(officeId);
+//        if (sessionQueue == null) {
+//            sessionQueue = new LinkedBlockingQueue<>();
+//            officeSessionMap.put(officeId, sessionQueue);
+//        }
+//        try {
+//            // Try to take an available session from the queue
+//            AmadeusSessionWrapper availableSession = sessionQueue.poll(2, TimeUnit.SECONDS);
+//            if (availableSession != null) {
+//                if (isValidSession(availableSession)) {
+//                    return availableSession;
+//                }
+//            }
+//            // Check the pool size
+//            if (sessionQueue.size() >= AmadeusConstants.SESSION_POOL_SIZE) {
+//                logger.debug("Amadeus session pooling max connection size reached. Waiting for connection...");
+//                throw new TimeoutException("Max session pool size reached, cannot create new sessions.");
+//            }
+//
+//            AmadeusSessionWrapper newSession = createSession(officeId);
+//            if (newSession != null) {
+//                sessionQueue.offer(newSession);
+//                return newSession;
+//            } else {
+//                throw new Exception("Failed to create a new session.");
+//            }
+//        } catch (InterruptedException ie) {
+//            logger.error("Thread was interrupted while waiting for a session.", ie);
+//        }
+//        return null;
+//    }
+
+    private boolean isValidSession(AmadeusSessionWrapper session) {
+        if (session.isQueryInProgress()) {
+            return false;
+        }
+
+        Period p = new Period(new DateTime(session.getLastQueryDate()), new DateTime(), PeriodType.minutes());
+        int inactivityTimeInMinutes = p.getMinutes();
+
+        if (inactivityTimeInMinutes >= AmadeusConstants.INACTIVITY_TIMEOUT) {
+            session.delete();
+            return false;
+        }
+
+        session.setQueryInProgress(true);
+        session.setLastQueryDate(new Date());
+        session.save();
+        return true;
+    }
+
+    public AmadeusSessionWrapper createSession(FlightSearchOffice office){
+        logger.debug("creating new  session .........................................office_id :" + office.getOfficeId());
+        try {
+            //ServiceHandler serviceHandler = new ServiceHandler();
+            AmadeusSessionWrapper amadeusSessionWrapper = serviceHandler.logIn(office);
+            return createSessionWrapper(amadeusSessionWrapper, office);
+        } catch (Exception e) {
+            logger.error("Amadeus createSession error " ,e);
+            e.printStackTrace();
+        }
+        return null;
+    }
+
+//    //todo
+    public AmadeusSessionWrapper getSession() throws Exception {
+        FlightSearchOffice office = sourceOfficeService.getAllOffices().get(0);
+        logger.debug("default officeId used in getSession :" + office.getOfficeId());
+        return getSession(office);
+    }
+
+    public synchronized AmadeusSessionWrapper getSession(FlightSearchOffice office) throws InterruptedException {
 
         logger.debug("AmadeusSessionManager getSession called");
        /* HashMap<String, AmadeusSessionWrapper> sessionHashMap = (HashMap<String, AmadeusSessionWrapper>) Cache.get(AmadeusConstants.AMADEUS_SESSION_LIST);
@@ -40,7 +117,7 @@ public class AmadeusSessionManager {
         if(sessionHashMap == null) {
             sessionHashMap = new HashMap<>();
         }*/
-        List<AmadeusSessionWrapper> amadeusSessionWrapperList = AmadeusSessionWrapper.findAllInactiveContextList();
+        List<AmadeusSessionWrapper> amadeusSessionWrapperList = AmadeusSessionWrapper.findAllInactiveContextListByOfficeId(office.getOfficeId());
         int count = 0;
         for(AmadeusSessionWrapper amadeusSessionWrapper : amadeusSessionWrapperList){
             count++;
@@ -57,37 +134,22 @@ public class AmadeusSessionManager {
             amadeusSessionWrapper.setLastQueryDate(new Date());
 //            updateAmadeusSession(amadeusSessionWrapper);
             amadeusSessionWrapper.save();
+            logger.debug("Returning existing session ........................................." + amadeusSessionWrapper.getmSession().value.getSessionId());
             System.out.println("Returning existing session ........................................." + amadeusSessionWrapper.getmSession().value.getSessionId());
             return amadeusSessionWrapper;
         }
         if(count >= AmadeusConstants.SESSION_POOL_SIZE){
             logger.debug("Amadeus session pooling max connection size reached waiting for connection...................");
             Thread.sleep(2000);
-            getSession();
+            getSession(office);
         }else {
-            AmadeusSessionWrapper amadeusSessionWrapper = createSession();
+            AmadeusSessionWrapper amadeusSessionWrapper = createSession(office);
 /*            sessionHashMap.put(amadeusSessionWrapper.getmSession().value.getSecurityToken(),amadeusSessionWrapper);
             Cache.set(AmadeusConstants.AMADEUS_SESSION_LIST, sessionHashMap);*/
             return amadeusSessionWrapper;
         }
-
         return null;
     }
-
-
-    public AmadeusSessionWrapper createSession(){
-        System.out.println("creating new  session .........................................");
-        try {
-            ServiceHandler serviceHandler = new ServiceHandler();
-            SessionHandler sessionHandler = serviceHandler.logIn(new SessionHandler());
-            return createSessionWrapper(sessionHandler.getSession().value);
-        } catch (Exception e) {
-            logger.error("Amadeus createSession error " ,e);
-            e.printStackTrace();
-        }
-        return null;
-    }
-
 
     public AmadeusSessionWrapper createSessionWrapper(Session session){
         AmadeusSessionWrapper amadeusSessionWrapper = new AmadeusSessionWrapper();
@@ -99,15 +161,45 @@ public class AmadeusSessionManager {
         return amadeusSessionWrapper;
     }
 
-    public void updateAmadeusSession(AmadeusSessionWrapper amadeusSessionWrapper){
+    public AmadeusSessionWrapper createSessionWrapper(AmadeusSessionWrapper amadeusSessionWrapper,FlightSearchOffice office){
+        //AmadeusSessionWrapper amadeusSessionWrapper = new AmadeusSessionWrapper();
+        amadeusSessionWrapper.setActiveContext(false);
         amadeusSessionWrapper.setQueryInProgress(false);
-        amadeusSessionWrapper.update();
+        amadeusSessionWrapper.setLastQueryDate(new Date());
+//        //amadeusSessionWrapper.setmSession(new Holder<>(session));
+//        amadeusSessionWrapper.setOfficeId(office.getGetOfficeId());
+//        if(office.isPartner()) {
+//            amadeusSessionWrapper.setPartnerName("Benji");
+//        }
+        amadeusSessionWrapper.save();
+        return amadeusSessionWrapper;
     }
 
-    public String storeActiveSession(Session session, String pnr){
+//    public AmadeusSessionWrapper createSessionWrapper(AmadeusSessionWrapper amadeusSessionWrapper,FlightSearchOffice office){
+//        //AmadeusSessionWrapper amadeusSessionWrapper = new AmadeusSessionWrapper();
+//        amadeusSessionWrapper.setActiveContext(false);
+//        amadeusSessionWrapper.setQueryInProgress(false);
+//        amadeusSessionWrapper.setLastQueryDate(new Date());
+//        //amadeusSessionWrapper.setmSession(new Holder<>(session));
+//        amadeusSessionWrapper.setOfficeId(office.getGetOfficeId());
+//        if(office.isPartner()) {
+//            amadeusSessionWrapper.setPartnerName("Benji");
+//        }
+//        amadeusSessionWrapper.save();
+//        return amadeusSessionWrapper;
+//    }
+
+    public void updateAmadeusSession(AmadeusSessionWrapper amadeusSessionWrapper){
+        if(amadeusSessionWrapper != null) {
+            amadeusSessionWrapper.setQueryInProgress(false);
+            amadeusSessionWrapper.update();
+        }
+    }
+
+    public String storeActiveSession(AmadeusSessionWrapper amadeusSessionWrapper, String pnr){
 
         String uuid = UUID.randomUUID().toString();
-        AmadeusSessionWrapper amadeusSessionWrapper = createSessionWrapper(session);
+        //AmadeusSessionWrapper amadeusSessionWrapper = createSessionWrapper(session);
         amadeusSessionWrapper.setSessionUUID(uuid);
         amadeusSessionWrapper.setActiveContext(true);
         amadeusSessionWrapper.setGdsPNR(pnr);
@@ -120,13 +212,25 @@ public class AmadeusSessionManager {
         return amadeusSessionWrapper.getmSession().value;
     }
 
-    public void removeActiveSession(Session session){
-        AmadeusSessionWrapper amadeusSessionWrapper = AmadeusSessionWrapper.findBySessionId(session.getSessionId());
-        amadeusSessionWrapper.delete();
+    public AmadeusSessionWrapper getActiveSessionByRef(String sessionIdRef){
+        AmadeusSessionWrapper amadeusSessionWrapper = AmadeusSessionWrapper.findSessionByUUID(sessionIdRef);
+        return amadeusSessionWrapper;
     }
 
-    public Session getActiveSessionByGdsPNR(String pnr){
+    //todo
+    public void removeActiveSession(Session session){
+        AmadeusSessionWrapper amadeusSessionWrapper = AmadeusSessionWrapper.findBySessionId(session.getSessionId());
+        if(amadeusSessionWrapper != null)
+            amadeusSessionWrapper.delete();
+    }
+
+//    public Session getActiveSessionByGdsPNR(String pnr){
+//        AmadeusSessionWrapper amadeusSessionWrapper = AmadeusSessionWrapper.findByPNR(pnr);
+//        return amadeusSessionWrapper.getmSession().value;
+//    }
+
+    public AmadeusSessionWrapper getActiveSessionByGdsPNR(String pnr){
         AmadeusSessionWrapper amadeusSessionWrapper = AmadeusSessionWrapper.findByPNR(pnr);
-        return amadeusSessionWrapper.getmSession().value;
+        return amadeusSessionWrapper;
     }
 }
